requirements.txt
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
numpy==1.24.3
scikit-learn==1.3.2
pillow==10.0.1
tqdm==4.66.1
matplotlib==3.8.2
pandas==2.1.3
